Add information about user state loading status

+++ src/app/slices/decksSlice.tsx
@@ -1,5 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { setSuccess, setGenericError } from './messageSlice';
+import { setSuccess, setGenericError, setError } from './messageSlice';
 import deckService from '../services/deckService';
 import userService from '../services/userService';
 import type { PayloadAction } from '@reduxjs/toolkit';
@@ -22,19 +22,19 @@ export const decksSlice = createSlice({
       return action.payload;
     },
     appendDeck: (state, action: PayloadAction<IDeck>) => {
-      if (state.owned === null || state.owned === undefined) {
+      if (state.owned == null) {
         return { owned: [action.payload] };
       }
       state.owned.push(action.payload);
     },
     replaceDeck: (state, action: PayloadAction<IDeck>) => {
-      if (state.owned === null || state.owned === undefined) {
+      if (state.owned == null) {
         return state;
       }
       return { owned: state.owned.map(deck => deck.id === action.payload.id ? action.payload : deck) };
     },
     removeDeck: (state, action: PayloadAction<IDeck>) => {
-      if (state.owned === null || state.owned === undefined) {
+      if (state.owned == null) {
         return state;
       }
       return { owned: state.owned.filter(deck => deck.id !== action.payload.id) };
@@ -47,8 +47,8 @@ export const { setDecks, appendDeck, replaceDeck } = decksSlice.actions;
 export const loadDecks = () => {
   return async (dispatch: AppDispatch, getState: AppGetState) => {
     const state = getState();
-    if (state.user.id !== 0) {
-      const response = await deckService.getUserDecks(state.user.id, state.user.token);
+    if (state.user.self != null && state.user.token != null) {
+      const response = await deckService.getUserDecks(state.user.self.id, state.user.token);
       dispatch(setDecks({ owned: response.data }));
     } else {
       dispatch(setDecks(initialState));
@@ -59,8 +59,12 @@ export const loadDecks = () => {
 export const createDeck = (name: string) => {
   return async (dispatch: AppDispatch, getState: AppGetState) => {
     const state = getState();
-    const response = await deckService.postDeck(name, state.user.token);
-    dispatch(appendDeck(response.data));
+    if (state.user.token != null) {
+      const response = await deckService.postDeck(name, state.user.token);
+      dispatch(appendDeck(response.data));
+    } else {
+      dispatch(setError('You are not logged in'));
+    }
   };
 };
 
@@ -68,9 +72,13 @@ export const updateDeck = (deck: IDeck) => {
   return async (dispatch: AppDispatch, getState: AppGetState) => {
     const state = getState();
     try {
-      const response = await deckService.updateDeck(deck, state.user.token);
-      dispatch(replaceDeck(response.data));
-      dispatch(setSuccess('Deck updated successfully'));
+      if (state.user.token != null) {
+        const response = await deckService.updateDeck(deck, state.user.token);
+        dispatch(replaceDeck(response.data));
+        dispatch(setSuccess('Deck updated successfully'));
+      } else {
+        dispatch(setError('You are not logged in'));
+      }
     } catch (e) {
       void dispatch(setGenericError(e));
     }
@@ -81,9 +89,13 @@ export const deleteDeck = (deckId: number) => {
   return async (dispatch: AppDispatch, getState: AppGetState) => {
     const state = getState();
     try {
-      const response = await deckService.deleteDeck(deckId, state.user.token);
-      dispatch(decksSlice.actions.removeDeck(response.data));
-      dispatch(setSuccess('Deck deleted successfully'));
+      if (state.user.token != null) {
+        const response = await deckService.deleteDeck(deckId, state.user.token);
+        dispatch(decksSlice.actions.removeDeck(response.data));
+        dispatch(setSuccess('Deck deleted successfully'));
+      } else {
+        dispatch(setError('You are not logged in'));
+      }
     } catch (e) {
       void dispatch(setGenericError(e));
     }
@@ -94,9 +106,13 @@ export const duplicateDeck = (deckId: number) => {
   return async (dispatch: AppDispatch, getState: AppGetState) => {
     const state = getState();
     try {
-      const response = await userService.addDeck(state.user.id, deckId, state.user.token);
-      dispatch(appendDeck(response.data));
-      dispatch(setSuccess('Deck cloned successfully'));
+      if (state.user.self != null && state.user.token != null) {
+        const response = await userService.addDeck(state.user.self.id, deckId, state.user.token);
+        dispatch(appendDeck(response.data));
+        dispatch(setSuccess('Deck cloned successfully'));
+      } else {
+        dispatch(setError('You are not logged in'));
+      }
     } catch (e) {
       void dispatch(setGenericError(e));
     }
